cyber-copilot/requirements.txt
flask>=3.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
openai>=1.50.0
anthropic>=0.34.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
rich>=13.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
